Assignment 1 Continue working with the project from the previous homework. For the product model, implement the CRUD mechanism using the module django.forms .
Conditions for users:

 can create new products;
 cannot upload forbidden products to the platform.
To exclude uploading prohibited products, implement validation of the product name and description in such a way that words cannot be added to them: casino, cryptocurrency, crypta, exchange, cheap, free, scam, police, radar.

Task 2
Add a new model "Version" which should contain the following fields:

product,
version number,
version name,
sign of the current version.
If there is an active version, implement the output of information about the active version to the list of products.

Task 3
To work with product versions, add an implementation of working with forms. A version can only be added to an existing product.

All created forms should be styled in such a way that they are in the same design style of the whole platform. To do this, you can use the method
__init__
 or study the crispy-forms package yourself: https://pypi.org/project/django-crispy-forms/.